Create a FeeManager for Uniswap V3
This PR adds a contract to manage protocol fee acquisition for Uniswap V3. If this FeeManager contract were made the Owner of the UniswapV3Factory, it would allow for the collection and eventual distribution of protocol fees to token holders. The PR also adds a basic script to facilitate deployment of the FeeManager.